pyproject.toml
[tool.poetry]
name = "notion-google-calendar"
version = "0.1.0"
description = ""
authors = ["CasperTeirlinck"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
google-api-python-client = "^2.81.0"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^1.0.0"
requests = "^2.28.2"
python-dotenv = "^1.0.0"
pyyaml = "^6.0"

[tool.poetry.group.dev.dependencies]
google-api-python-client-stubs = "^1.16.0"
black = "^23.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
